add: [ArchivedWarikanGroup]のmapToData extensionも追加
