--- Roulette/Model/UseCase/Extension/SeisanListExtension.swift
@@ -1,20 +1,0 @@
-//
-//  SeisanListExtension.swift
-//  Roulette
-//  
-//  Created by Seigetsu on 2023/12/21
-//  
-//
-
-import Foundation
-
-extension [Seisan] {
-    func mapToData(withMemberRepository memberRepository: MemberRepositoryProtocol) async throws -> [SeisanData] {
-        var seisanDataList: [SeisanData] = []
-        for seisan in self {
-            let seisanData = try await SeisanData.create(from: seisan, memberRepository: memberRepository)
-            seisanDataList.append(seisanData)
-        }
-        return seisanDataList
-    }
-}
